Upload a CSV manifest. Story #45
* Uses carrierwave gem (which is already a dependecy of Hyrax) to upload
and store the CSV manifest file.

* Uses environment variables to configure where the uploaded CSV files
will be stored, but the env vars are optional.  If not set, by default
it will use Hyrax's configured upload and cache directories.

* You must be logged in as an admin user to start a CSV-based import.

* When the user uploads a CSV file, we perform some basic validations on
the file and display a preview page that shows errors and warnings.

* After the user starts the import, the show page for a CsvImport will
display the current status of the import.  That is future work.  It
currently just has a placeholder in the show view.

* The index view for CsvImport records is just a placeholder fow now.
Fleshing out that page will be future work.

* I didn't add any navigation links to get to the new UI.  For now it's
a secret route.  We'll add menu links in a future PR.