Update on import (#213)
* Update on import

Darlingtonia 3.1 allows for a deduplication_field to be defined. Upon
import, darlingtonia will query for any records that already have a
given deduplication field (default is `identifier`). If it finds a
record that matches that field, it will update the metadata. If it does
not find an existing record, it will create a new record.

For now, updating the metadata of a record will not update the files
that are attached.