Remove navigation steps from create work spec
The navigation portions of the test rely on capybara matcher that may
timeout when loading the intermediate dashboard pages that occurr
in the user navigation flow before arriving at the new work form.

This assumes that the navigation to the page is not a primary concern
of the test and just loads the form view immediately.