Add a feature flipper to switch between new and old UI
We want to begin experimenting with breaking changes to user flows in the UI.
And we want to be able to easily compare changes with the current state and
fallback if required.  Therefore we're introducing a feature flipper and
wrapping UI additions, changes, and deletion in a toggle.

This change also cleans up indentation in the the _task partial.