Add a toggle-able batch operations group to the dashboard menu (#374)
We're trying to improve usability by making functions related to
batch operations easier to find.  Rather than having to navigate
to the main import page and then click to documentation or navigate
to a status page from there, this change gives direct links to those
functions in the left navigation bar in the dashboard.